Support for masked lt fringe + lt main kernels for col-major C matrix
Details:
- Updated the transposeGenerator and post-ops modules to support
  generating of lt kernels with >1 mask when C matrix is col-major.
- Support for row-major C matrix is yet to be done in post-ops.
- Removed standalone avx2 and avx512 gemm generators.

AMD-Internal: CPUPL-6959

+++ src/jit/amdzen/amdzen_generator.cc
@@ -671,7 +671,7 @@ jitAmdZenFP32::executeKernel(dlp::kernels::kernelParams* _params)
         }
 
         // Deploy the associated kernel
-        int kernel_idx = params->n_left;
+        int kernel_idx = static_cast<int>(params->n_left);
 
         utils::jit_gemv_m1_kernel kernel =
             reinterpret_cast<utils::jit_gemv_m1_kernel>(
